+++ packages/core-mobile/app/services/bridge/UnifiedBridgeService.ts
@@ -7,31 +7,20 @@ import {
   getEnabledBridgeServices,
   EvmSigner,
   BtcSigner,
-  Hex,
   BridgeAsset,
   Chain,
   TokenType,
-  isNativeAsset,
-  isErc20Asset,
   AnalyzeTxResult,
   AnalyzeTxParams,
   BridgeInitializer,
   GasSettings
 } from '@avalabs/bridge-unified'
-import { getBitcoinProvider } from 'services/network/utils/providerUtils'
 import { Network } from '@avalabs/core-chains-sdk'
 import { assertNotUndefined } from 'utils/assertions'
 import Logger from 'utils/Logger'
-import { TransactionParams } from '@avalabs/evm-module'
-import { AppListenerEffectAPI, RootState } from 'store'
-import { createInAppRequest } from 'store/rpc/utils/createInAppRequest'
-import { RpcMethod } from '@avalabs/vm-module-types'
-import { getBitcoinCaip2ChainId, getEvmCaip2ChainId } from 'temp/caip2ChainIds'
 import { addNamespaceToChain } from 'services/walletconnectv2/utils'
 import { BitcoinProvider } from '@avalabs/core-wallets-sdk'
-import { selectIsFeatureBlocked } from 'store/posthog'
-import { FeatureGates } from 'services/posthog/types'
-import { selectIsDeveloperMode } from 'store/settings/advanced'
+import { lowerCaseKeys } from 'utils/lowerCaseKeys'
 
 type BridgeService = ReturnType<typeof createUnifiedBridgeService>
 
@@ -40,59 +29,21 @@ export class UnifiedBridgeService {
 
   // init and fetch configs
   async init({
-    listenerApi
+    enabledBridgeTypes,
+    evmSigner,
+    btcSigner,
+    bitcoinProvider,
+    environment
   }: {
-    listenerApi: AppListenerEffectAPI
-  }): Promise<boolean> {
-    const state = listenerApi.getState()
-
-    if (this.isInitialized()) {
-      const prevState = listenerApi.getOriginalState()
-
-      if (!this.shouldReinitializeBridge(prevState, state)) return false
-    }
-
-    const featureGates = this.getFeatureGateStates(state)
-    const isTest = featureGates.isDeveloperMode
-    const enabledBridgeTypes = this.getEnabledBridgeTypes(featureGates)
-
-    const environment = isTest ? Environment.TEST : Environment.PROD
-    const request = createInAppRequest(listenerApi.dispatch)
-
+    enabledBridgeTypes: BridgeType[]
+    evmSigner: EvmSigner
+    btcSigner: BtcSigner
+    bitcoinProvider: BitcoinProvider
+    environment: Environment
+  }): Promise<void> {
     Logger.info('initializing unified bridge service', {
       environment
     })
-
-    const bitcoinProvider = await getBitcoinProvider(isTest)
-
-    const evmSigner: EvmSigner = {
-      sign: async ({ data, from, to, chainId }) => {
-        if (typeof to !== 'string') throw new Error('invalid to field')
-        const txParams: [TransactionParams] = [
-          {
-            from,
-            to,
-            data: data ?? undefined
-          }
-        ]
-
-        return request({
-          method: RpcMethod.ETH_SEND_TRANSACTION,
-          params: txParams,
-          chainId: getEvmCaip2ChainId(Number(chainId))
-        }) as Promise<Hex>
-      }
-    }
-
-    const btcSigner: BtcSigner = {
-      sign: async txData => {
-        return request({
-          method: RpcMethod.BITCOIN_SIGN_TRANSACTION,
-          params: txData,
-          chainId: getBitcoinCaip2ChainId(!isTest)
-        })
-      }
-    }
 
     const initializers = this.getBridgeInitializers({
       enabledBridgeTypes,
@@ -110,8 +61,6 @@ export class UnifiedBridgeService {
       environment,
       enabledBridgeServices
     })
-
-    return true
   }
 
   analyzeTx(params: AnalyzeTxParams): AnalyzeTxResult {
@@ -133,22 +82,23 @@ export class UnifiedBridgeService {
     sourceNetwork: Network
     targetNetwork: Network
   }): Promise<bigint> {
-    const feeMap = await this.service.getFees({
-      asset,
-      amount,
-      targetChain: await this.buildChain(targetNetwork),
-      sourceChain: await this.buildChain(sourceNetwork)
-    })
+    const feeMap = lowerCaseKeys(
+      await this.service.getFees({
+        asset,
+        amount,
+        targetChain: await this.buildChain(targetNetwork),
+        sourceChain: await this.buildChain(sourceNetwork)
+      })
+    )
 
-    if (isNativeAsset(asset)) {
-      // todo: handle this properly
-      return 0n
-    } else if (isErc20Asset(asset) && asset.address) {
-      const address = asset.address.toLowerCase() as `0x${string}`
-      return feeMap[address] ?? 0n
-    } else {
+    const identifier =
+      asset.type === TokenType.NATIVE ? asset.symbol : asset.address
+
+    if (!identifier) {
       throw new Error('invalid asset')
     }
+
+    return feeMap[identifier.toLowerCase()] ?? 0n
   }
 
   async transfer({
@@ -249,7 +199,7 @@ export class UnifiedBridgeService {
     })
   }
 
-  private isInitialized(): boolean {
+  isInitialized(): boolean {
     return this.#service !== undefined
   }
 
@@ -313,88 +263,6 @@ export class UnifiedBridgeService {
           }
       }
     })
-  }
-
-  // Check if any of the feature gate states or developer mode has changed
-  private shouldReinitializeBridge = (
-    prevState: RootState,
-    currState: RootState
-  ): boolean => {
-    const prevGates = this.getFeatureGateStates(prevState)
-    const currGates = this.getFeatureGateStates(currState)
-
-    return (
-      prevGates.isCctpBlocked !== currGates.isCctpBlocked ||
-      prevGates.isIcttBlocked !== currGates.isIcttBlocked ||
-      prevGates.isAbEvmBlocked !== currGates.isAbEvmBlocked ||
-      prevGates.isAbAvaToBtc !== currGates.isAbAvaToBtc ||
-      prevGates.isAbBtcToAva !== currGates.isAbBtcToAva ||
-      prevGates.isDeveloperMode !== currGates.isDeveloperMode
-    )
-  }
-
-  // Helper function to get the blocked status of feature gates
-  private getFeatureGateStates = (
-    state: RootState
-  ): {
-    isCctpBlocked: boolean
-    isIcttBlocked: boolean
-    isAbEvmBlocked: boolean
-    isAbAvaToBtc: boolean
-    isAbBtcToAva: boolean
-    isDeveloperMode: boolean
-  } => ({
-    isCctpBlocked: selectIsFeatureBlocked(
-      state,
-      FeatureGates.UNIFIED_BRIDGE_CCTP
-    ),
-    isIcttBlocked: selectIsFeatureBlocked(
-      state,
-      FeatureGates.UNIFIED_BRIDGE_ICTT
-    ),
-    isAbEvmBlocked: selectIsFeatureBlocked(
-      state,
-      FeatureGates.UNIFIED_BRIDGE_AB_EVM
-    ),
-    isAbAvaToBtc: selectIsFeatureBlocked(
-      state,
-      FeatureGates.UNIFIED_BRIDGE_AB_AVA_TO_BTC
-    ),
-    isAbBtcToAva: selectIsFeatureBlocked(
-      state,
-      FeatureGates.UNIFIED_BRIDGE_AB_BTC_TO_AVA
-    ),
-    isDeveloperMode: selectIsDeveloperMode(state)
-  })
-
-  // Get enabled bridge types based on feature gates
-  private getEnabledBridgeTypes = (featureGates: {
-    isCctpBlocked: boolean
-    isIcttBlocked: boolean
-    isAbEvmBlocked: boolean
-    isAbAvaToBtc: boolean
-    isAbBtcToAva: boolean
-    isDeveloperMode: boolean
-  }): BridgeType[] => {
-    const enabledBridgeTypes: BridgeType[] = []
-
-    if (!featureGates.isCctpBlocked) {
-      enabledBridgeTypes.push(BridgeType.CCTP)
-    }
-    if (!featureGates.isIcttBlocked) {
-      enabledBridgeTypes.push(BridgeType.ICTT_ERC20_ERC20)
-    }
-    if (!featureGates.isAbEvmBlocked) {
-      enabledBridgeTypes.push(BridgeType.AVALANCHE_EVM)
-    }
-    if (!featureGates.isAbAvaToBtc) {
-      enabledBridgeTypes.push(BridgeType.AVALANCHE_AVA_BTC)
-    }
-    if (!featureGates.isAbBtcToAva) {
-      enabledBridgeTypes.push(BridgeType.AVALANCHE_BTC_AVA)
-    }
-
-    return enabledBridgeTypes
   }
 }
