+++ packages/core-mobile/app/store/unifiedBridge/listeners.ts
@@ -1,13 +1,29 @@
-import { AppListenerEffectAPI } from 'store'
+import { AppListenerEffectAPI, RootState } from 'store'
 import { WalletState, onAppUnlocked, selectWalletState } from 'store/app'
 import { AppStartListening } from 'store/middleware/listener'
-import { toggleDeveloperMode } from 'store/settings/advanced'
+import {
+  selectIsDeveloperMode,
+  toggleDeveloperMode
+} from 'store/settings/advanced'
 import { isAnyOf } from '@reduxjs/toolkit'
 import UnifiedBridgeService from 'services/bridge/UnifiedBridgeService'
-import { BridgeTransfer } from '@avalabs/bridge-unified'
-import { setFeatureFlags } from 'store/posthog'
+import {
+  BridgeTransfer,
+  BridgeType,
+  BtcSigner,
+  Environment,
+  EvmSigner,
+  Hex
+} from '@avalabs/bridge-unified'
+import { selectIsFeatureBlocked, setFeatureFlags } from 'store/posthog'
 import { showTransactionSuccessToast } from 'utils/toast'
 import Logger from 'utils/Logger'
+import { createInAppRequest } from 'store/rpc/utils/createInAppRequest'
+import { FeatureGates } from 'services/posthog/types'
+import { getBitcoinProvider } from 'services/network/utils/providerUtils'
+import { RpcMethod } from '@avalabs/vm-module-types'
+import { getBitcoinCaip2ChainId, getEvmCaip2ChainId } from 'temp/caip2ChainIds'
+import { TransactionParams } from '@avalabs/evm-module'
 import {
   removePendingTransfer,
   selectPendingTransfers,
@@ -62,13 +78,59 @@ export const initUnifiedBridgeService = async (
   // No need to proceed if the wallet is not unlocked
   if (walletState !== WalletState.ACTIVE) return
 
-  const initialized = await UnifiedBridgeService.init({
-    listenerApi
+  const request = createInAppRequest(listenerApi.dispatch)
+
+  if (UnifiedBridgeService.isInitialized()) {
+    const prevState = listenerApi.getOriginalState()
+
+    if (!shouldReinitializeBridge(prevState, state)) return
+  }
+
+  const featureGates = getFeatureGateStates(state)
+  const enabledBridgeTypes = getEnabledBridgeTypes(featureGates)
+  const isTest = featureGates.isDeveloperMode
+
+  const bitcoinProvider = await getBitcoinProvider(isTest)
+
+  const evmSigner: EvmSigner = {
+    sign: async ({ data, from, to, chainId }) => {
+      if (typeof to !== 'string') throw new Error('invalid to field')
+      const txParams: [TransactionParams] = [
+        {
+          from,
+          to,
+          data: data ?? undefined
+        }
+      ]
+
+      return request({
+        method: RpcMethod.ETH_SEND_TRANSACTION,
+        params: txParams,
+        chainId: getEvmCaip2ChainId(Number(chainId))
+      }) as Promise<Hex>
+    }
+  }
+
+  const btcSigner: BtcSigner = {
+    sign: async txData => {
+      return request({
+        method: RpcMethod.BITCOIN_SIGN_TRANSACTION,
+        params: txData,
+        chainId: getBitcoinCaip2ChainId(!isTest)
+      })
+    }
+  }
+  const environment = isTest ? Environment.TEST : Environment.PROD
+
+  await UnifiedBridgeService.init({
+    enabledBridgeTypes,
+    evmSigner,
+    btcSigner,
+    bitcoinProvider,
+    environment
   })
 
-  if (initialized) {
-    trackPendingTransfers(listenerApi)
-  }
+  trackPendingTransfers(listenerApi)
 }
 
 const checkTransferStatus = async (
@@ -99,4 +161,86 @@ export const addUnifiedBridgeListeners = (
     actionCreator: setPendingTransfer,
     effect: checkTransferStatus
   })
+}
+
+// Helper function to get the blocked status of feature gates
+const getFeatureGateStates = (
+  state: RootState
+): {
+  isCctpBlocked: boolean
+  isIcttBlocked: boolean
+  isAbEvmBlocked: boolean
+  isAbAvaToBtc: boolean
+  isAbBtcToAva: boolean
+  isDeveloperMode: boolean
+} => ({
+  isCctpBlocked: selectIsFeatureBlocked(
+    state,
+    FeatureGates.UNIFIED_BRIDGE_CCTP
+  ),
+  isIcttBlocked: selectIsFeatureBlocked(
+    state,
+    FeatureGates.UNIFIED_BRIDGE_ICTT
+  ),
+  isAbEvmBlocked: selectIsFeatureBlocked(
+    state,
+    FeatureGates.UNIFIED_BRIDGE_AB_EVM
+  ),
+  isAbAvaToBtc: selectIsFeatureBlocked(
+    state,
+    FeatureGates.UNIFIED_BRIDGE_AB_AVA_TO_BTC
+  ),
+  isAbBtcToAva: selectIsFeatureBlocked(
+    state,
+    FeatureGates.UNIFIED_BRIDGE_AB_BTC_TO_AVA
+  ),
+  isDeveloperMode: selectIsDeveloperMode(state)
+})
+
+// Check if any of the feature gate states or developer mode has changed
+const shouldReinitializeBridge = (
+  prevState: RootState,
+  currState: RootState
+): boolean => {
+  const prevGates = getFeatureGateStates(prevState)
+  const currGates = getFeatureGateStates(currState)
+
+  return (
+    prevGates.isCctpBlocked !== currGates.isCctpBlocked ||
+    prevGates.isIcttBlocked !== currGates.isIcttBlocked ||
+    prevGates.isAbEvmBlocked !== currGates.isAbEvmBlocked ||
+    prevGates.isAbAvaToBtc !== currGates.isAbAvaToBtc ||
+    prevGates.isAbBtcToAva !== currGates.isAbBtcToAva ||
+    prevGates.isDeveloperMode !== currGates.isDeveloperMode
+  )
+}
+
+// Get enabled bridge types based on feature gates
+const getEnabledBridgeTypes = (featureGates: {
+  isCctpBlocked: boolean
+  isIcttBlocked: boolean
+  isAbEvmBlocked: boolean
+  isAbAvaToBtc: boolean
+  isAbBtcToAva: boolean
+  isDeveloperMode: boolean
+}): BridgeType[] => {
+  const enabledBridgeTypes: BridgeType[] = []
+
+  if (!featureGates.isCctpBlocked) {
+    enabledBridgeTypes.push(BridgeType.CCTP)
+  }
+  if (!featureGates.isIcttBlocked) {
+    enabledBridgeTypes.push(BridgeType.ICTT_ERC20_ERC20)
+  }
+  if (!featureGates.isAbEvmBlocked) {
+    enabledBridgeTypes.push(BridgeType.AVALANCHE_EVM)
+  }
+  if (!featureGates.isAbAvaToBtc) {
+    enabledBridgeTypes.push(BridgeType.AVALANCHE_AVA_BTC)
+  }
+  if (!featureGates.isAbBtcToAva) {
+    enabledBridgeTypes.push(BridgeType.AVALANCHE_BTC_AVA)
+  }
+
+  return enabledBridgeTypes
 }
