+++ packages/core-mobile/app/utils/lowerCaseKeys.ts
@@ -1,0 +1,5 @@
+export const lowerCaseKeys = <T>(obj: Record<string, T>): Record<string, T> => {
+  return Object.fromEntries(
+    Object.entries(obj).map(([key, value]) => [key.toLowerCase(), value])
+  )
+}
